+++ driver/src/main/scala/org/mongodb/scala/model/package.scala
@@ -17,7 +17,6 @@
 package org.mongodb.scala
 
 import scala.collection.JavaConverters._
-
 import org.mongodb.scala.bson.conversions.Bson
 
 // scalastyle:off number.of.methods number.of.types
@@ -77,22 +76,6 @@ package object model {
    */
   object BucketOptions {
     def apply(): BucketOptions = new com.mongodb.client.model.BucketOptions()
-  }
-
-  /**
-   * The options for a \$out aggregation pipeline stage
-   *
-   * @since 2.7
-   */
-  type AggregateOutStageOptions = com.mongodb.client.model.AggregateOutStageOptions
-
-  /**
-   * The options for a \$out aggregation pipeline stage
-   *
-   * @since 2.7
-   */
-  object AggregateOutStageOptions {
-    def apply(): AggregateOutStageOptions = new com.mongodb.client.model.AggregateOutStageOptions()
   }
 
   /**