+++ docs/reference/content/changelog.md
@@ -13,6 +13,7 @@ Changes between released versions
 ### 2.7.0
   * Updated MongoDB Driver Async to 3.11.0
   * Updated DEFAULT_CODEC_REGISTRY order, as upstream BsonCodec captures all Bson instances and priority needs to be lower.
+  * Support $merge pipeline stage in aggregations [SCALA-473](https://jira.mongodb.org/browse/SCALA-473)
 
 ### 2.6.0
   * Updated MongoDB Driver Async to 3.10.0